Обработчики сообщений вынесены в отдельный файл
